Cleanup: move bmqsys_statmonitor -> mqbstat_statmonitor
