--- src/groups/bmq/bmqsys/doc/bmqsys.txt
@@ -1,21 +1,0 @@
-    bmqsys.txt
-
-@PURPOSE: Provide systems utilities and mechanisms.
-
-@MNEMONIC: BlazingMQ Systems (bmqsys)
-
-@DESCRIPTION: This package provides miscellaneous systems related utility
-components to be reused through various applications.
-
-/Hierarchical Synopsis
-/---------------------
-The 'bmqsys' package currently has 1 component having 1 level of physical
-dependency.  The list below shows the hierarchal ordering of the components.
-..
-  1. bmqsys_statmonitor
-..
-
-/Component Synopsis
-/------------------
-: 'bmqsys_statmonitor':
-:      Provide a mechanism to monitor cpu and memory stats.
