--- src/groups/bmq/bmqsys/package/bmqsys.dep
@@ -1,3 +1,0 @@
-bmqu
-bmqscm
-bmqst
