--- src/groups/bmq/bmqsys/package/bmqsys.mem
@@ -1,1 +1,0 @@
-bmqsys_statmonitor
