+++ src/groups/bmq/group/bmq.mem
@@ -12,7 +12,6 @@ bmqpi
 bmqscm
 bmqst
 bmqstm
-bmqsys
 bmqt
 bmqtsk
 bmqtst
