--- src/groups/bmq/group/mwc.mem
@@ -1,12 +1,0 @@
-bmqc
-bmqex
-bmqio
-bmqma
-bmqscm
-bmqst
-bmqstm
-bmqsys
-bmqt
-bmqtsk
-bmqtst
-bmqu
