+++ src/groups/mqb/mqbstat/mqbstat_statcontroller.h
@@ -27,14 +27,13 @@
 // config) dumping the stats to a dedicated log file.
 
 // MQB
-
 #include <mqbcmd_messages.h>
 #include <mqbstat_jsonprinter.h>
 #include <mqbstat_printer.h>
+#include <mqbstat_statmonitor.h>
 
 #include <bmqma_countingallocatorstore.h>
 #include <bmqst_statcontext.h>
-#include <bmqsys_statmonitor.h>
 
 // BDE
 #include <ball_log.h>
@@ -126,7 +125,7 @@ class StatController {
     typedef bslma::ManagedPtr<bdlmt::TimerEventScheduler> SchedulerMp;
     typedef bslma::ManagedPtr<bmqst::StatContext>         StatContextMp;
     typedef bsl::shared_ptr<bmqst::StatContext>           StatContextSp;
-    typedef bslma::ManagedPtr<bmqsys::StatMonitor>        SystemStatMonitorMp;
+    typedef bslma::ManagedPtr<mqbstat::StatMonitor>       SystemStatMonitorMp;
     typedef bslma::ManagedPtr<Printer>                    PrinterMp;
     typedef bslma::ManagedPtr<JsonPrinter>                JsonPrinterMp;
     typedef bslma::ManagedPtr<mqbplug::StatPublisher>     StatPublisherMp;
